+++ api/src/main/java/ca/bc/gov/educ/api/psi/repository/PsiCriteriaQueryRepository.java
@@ -3,7 +3,7 @@ package ca.bc.gov.educ.api.psi.repository;
 import org.springframework.stereotype.Repository;
 
 import ca.bc.gov.educ.api.psi.model.entity.PsiEntity;
-import ca.bc.gov.educ.api.psi.repository.criteria.CriteriaQueryRepository;
+import ca.bc.gov.educ.api.psi.util.criteria.CriteriaQueryRepository;
 
 @Repository
 public interface PsiCriteriaQueryRepository extends CriteriaQueryRepository<PsiEntity> {
