+++ api/src/main/java/ca/bc/gov/educ/api/psi/repository/PsiCriteriaQueryRepositoryImpl.java
@@ -3,7 +3,7 @@ package ca.bc.gov.educ.api.psi.repository;
 import org.springframework.stereotype.Repository;
 
 import ca.bc.gov.educ.api.psi.model.entity.PsiEntity;
-import ca.bc.gov.educ.api.psi.repository.criteria.CriteriaQueryRepositoryImpl;
+import ca.bc.gov.educ.api.psi.util.criteria.CriteriaQueryRepositoryImpl;
 
 @Repository
 public class PsiCriteriaQueryRepositoryImpl extends CriteriaQueryRepositoryImpl<PsiEntity> implements PsiCriteriaQueryRepository {
