+++ api/src/main/java/ca/bc/gov/educ/api/psi/util/criteria/CriteriaHelper.java
@@ -1,4 +1,4 @@
-package ca.bc.gov.educ.api.psi.repository.criteria;
+package ca.bc.gov.educ.api.psi.util.criteria;
 
 import java.util.ArrayList;
 import java.util.Collection;
@@ -9,7 +9,7 @@ import java.util.Map;
 import org.apache.commons.lang3.StringUtils;
 import org.jboss.logging.Logger;
 
-import ca.bc.gov.educ.api.psi.repository.criteria.GradCriteria.OperationEnum;
+import ca.bc.gov.educ.api.psi.util.criteria.GradCriteria.OperationEnum;
 
 public class CriteriaHelper {
